+++ commandapi-kotlin/src/main/kotlin/dev/jorel/commandapi/kotlindsl/CommandTreeDSL.kt
@@ -1,5 +1,6 @@
-package dev.jorel.commandapi
+package dev.jorel.commandapi.kotlindsl
 
+import dev.jorel.commandapi.*
 import dev.jorel.commandapi.arguments.*
 import dev.jorel.commandapi.arguments.LiteralArgument.of
 import dev.jorel.commandapi.arguments.ScoreHolderArgument.ScoreHolderType
